PR suggestion of further nesting structure to allow interface specific util functions
