+++ nwb_conversion_tools/datainterfaces/behavior/movie/moviedatainterface.py
@@ -11,9 +11,9 @@ from pynwb.image import ImageSeries
 from hdmf.backends.hdf5.h5_utils import H5DataIO
 from hdmf.data_utils import DataChunkIterator
 
-from ...basedatainterface import BaseDataInterface
-from ..interface_utils.movie_utils import get_movie_timestamps, get_movie_fps, get_frame_shape
-from ...utils.conversion_tools import check_regular_timestamps
+from ....basedatainterface import BaseDataInterface
+from .movie_utils import get_movie_timestamps, get_movie_fps, get_frame_shape
+from ....utils.conversion_tools import check_regular_timestamps
 
 try:
     import cv2
