+++ nwb_conversion_tools/datainterfaces/ecephys/axona/axonadatainterface.py
@@ -10,10 +10,10 @@ from pynwb import NWBFile
 from pynwb.behavior import Position, SpatialSeries
 from pynwb.ecephys import ElectricalSeries
 
-from ...utils.json_schema import get_schema_from_hdmf_class
-from ...basedatainterface import BaseDataInterface
-from .baserecordingextractorinterface import BaseRecordingExtractorInterface
-from ...utils.conversion_tools import get_module
+from ....utils.json_schema import get_schema_from_hdmf_class
+from ....basedatainterface import BaseDataInterface
+from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
+from ....utils.conversion_tools import get_module
 
 
 # Helper functions for AxonaRecordingExtractorInterface
