+++ nwb_conversion_tools/datainterfaces/ecephys/tutorial/ecephystutorialdatainterface.py
@@ -4,9 +4,9 @@ from pathlib import Path
 import numpy as np
 import spikeextractors as se
 
-from ...utils.json_schema import get_schema_from_method_signature
-from .baserecordingextractorinterface import BaseRecordingExtractorInterface
-from .basesortingextractorinterface import BaseSortingExtractorInterface
+from ....utils.json_schema import get_schema_from_method_signature
+from ..baserecordingextractorinterface import BaseRecordingExtractorInterface
+from ..basesortingextractorinterface import BaseSortingExtractorInterface
 
 
 class TutorialRecordingInterface(BaseRecordingExtractorInterface):
